refactor: changed to another package the classes
