+++ src/main/java/br/com/matheus/school/domain/student/CPF.java
@@ -1,4 +1,4 @@
-package br.com.matheus;
+package br.com.matheus.school.domain.student;
 
 public class CPF {
