+++ src/main/java/br/com/matheus/school/domain/student/FactoryStudent.java
@@ -1,4 +1,4 @@
-package br.com.matheus;
+package br.com.matheus.school.domain.student;
 
 public class FactoryStudent {
