+++ src/test/java/br/com/matheus/school/fake/FakeCPF.java
@@ -1,6 +1,6 @@
-package br.com.matheus.fake;
+package br.com.matheus.school.fake;
 
-import br.com.matheus.CPF;
+import br.com.matheus.school.domain.student.CPF;
 
 public class FakeCPF {
