+++ src/test/java/br/com/matheus/school/fake/FakePhone.java
@@ -1,6 +1,6 @@
-package br.com.matheus.fake;
+package br.com.matheus.school.fake;
 
-import br.com.matheus.Phone;
+import br.com.matheus.school.domain.student.Phone;
 
 public class FakePhone {
